Pt 4
trying again to push changes: Pt 4

While there are significant changes to the manuscript and readme files with this update, there are still sections that need work. The EDA code section needs to organized and the analysis code will be augmented. A better data source was identified and requires more processing. THis will all be completed in Pt 5.